requirements.txt
streamlit
gspread
pillow
